requirements.txt
fastapi
langchain
langchain-openai
langchain-core
langchain-community
supabase
dotenv
uvicorn
python-jose
bcrypt
jwt
